+++ src/test/java/com/dokkebi/officefinder/repository/ReviewRepositoryTest.java
@@ -5,6 +5,7 @@ import com.dokkebi.officefinder.entity.lease.Lease;
 import com.dokkebi.officefinder.entity.office.Office;
 import com.dokkebi.officefinder.entity.review.Review;
 import com.dokkebi.officefinder.entity.type.LeaseStatus;
+import com.dokkebi.officefinder.repository.lease.LeaseRepository;
 import com.dokkebi.officefinder.repository.office.OfficeRepository;
 import java.util.Set;
 import org.junit.jupiter.api.Assertions;
